Hide TimeRef
