--- router/time_ref.ts
@@ -1,22 +1,0 @@
-export class TimeRef<T>
-{
-  private _ref: T | undefined
-
-  constructor(ref: T, delay: number = 6e4, block?: (ref: T) => boolean | void)
-  {
-    this._ref = ref
-
-    setTimeout(() => {
-      if (this._ref) {
-        if (!(block && block(this._ref))) {
-          this._ref = undefined
-        }
-      }
-    }, delay)
-  }
-
-  public deref()
-  {
-    return this._ref
-  }
-}
